App/Code/requirement.txt
kafka-python==2.0.2
psycopg2
requests
uuid
six